# Source Code/wwwroot/api_documentation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoneyTrackr API",
    "description": "Sample API to manage deposits and withdrawals entries, it also implements User management. All endpoints except the 'Login' and 'Register' requires the 'Authorization' header to be set using the JWT provided by the 'Login' endpoint preceded by the 'Bearer' indicator.",
    "contact": {
      "name": "Rolando Malena",
      "url": "https://www.linkedin.com/in/rolandomalena/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/v1/Access/Register": {
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "Register a new User. Will be given a Regular User role.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/RegisterUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Registered successfully."
          },
          "400": {
            "description": "One or more validations error encountered"
          }
        }
      }
    },
    "/api/v1/Access/Login": {
      "post": {
        "tags": [
          "Access"
        ],
        "summary": "Logs in User and returns a JWT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/LogInUserRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Logged In successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/AuthorizationTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Username or Password is incorrect or User has been blocked out."
          }
        }
      }
    },
...
    "/api/v1/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get every User registered",
        "responses": {
          "200": {
            "description": "User List retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/UserRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "One or more validations error encountered"
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Users/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets a single User by its Username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The Username of the user to be found",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSchema"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          },
          "404": {
            "description": "User was not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Updates a specific User",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/requestBodies/UserRequestBody"
              }
            }
          },
          "description": "NOTE: Only the provided fields will be updated."
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "One or more validations error encountered"
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          },
          "404": {
            "description": "User was not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a specific User.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": false
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "400": {
            "description": "User tried to delete him/herself"
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          },
          "404": {
            "description": "User was not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/Roles/{roleId}/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get every User registered by its Role",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User List retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/UsersResponse"
                }
              }
            }
          },
          "400": {
            "description": "The provided roleId is not valid"
          },
          "401": {
            "description": "Unauthorized. Use a Bearer Token"
          },
          "403": {
            "description": "User has no access to this resource"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\". If you don't have a user account, you can create one using the 'Register' method. You can get your JWT using the 'Login' method.",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "RoleSchema": {
        "title": "Role",
        "type": "object",
        "properties": {
          "id": {
            "description": "Role Identifier",
            "type": "string",
            "nullable": false
          },
          "name": {
            "description": "Role Name",
            "type": "string",
            "nullable": false
          }
        }
      },
      "UserSchema": {
        "title": "User",
        "type": "object",
        "properties": {
          "id": {
            "description": "User Identifier",
            "type": "string",
            "nullable": false
          },
          "userName": {
            "description": "User Name",
            "type": "string",
            "nullable": false
          },
          "role": {
            "description": "User Role",
            "type": "object",
            "nullable": false,
            "$ref": "#/components/schemas/RoleSchema"
          }
        }
      },
...
    },
    "requestBodies": {
      "RegisterUserRequestBody": {
        "title": "Register User Request Body",
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 25,
            "minLength": 5,
            "type": "string",
            "nullable": false
          },
          "password": {
            "maxLength": 25,
            "minLength": 5,
            "type": "string",
            "example": "password1",
            "nullable": false,
            "description": "NOTE: Passwords must contain at least one alphabetic character and at least one numeric character."
          }
        }
      },
      "LogInUserRequestBody": {
        "title": "Log In User Request Body",
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false,
            "example": "password1"
          }
        }
      },
      "UserRequestBody": {
        "title": "User Request Body",
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "maxLength": 25,
            "minLength": 5,
            "type": "string",
            "nullable": true
          },
          "password": {
            "maxLength": 25,
            "minLength": 5,
            "type": "string",
            "example": "password1",
            "nullable": true,
            "description": "NOTE: Passwords must contain at least one alphabetic character and at least one numeric character."
          }
        }
      },
...
    },
    "responses": {
      "AuthorizationTokenResponse": {
        "title": "Authorization Token Response Object",
        "type": "object",
        "properties": {
          "auth_token": {
            "description": "Bearer Token to be used in all future requests.",
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "UsersResponse": {
        "title": "Get Users Response",
        "type": "array",
        "$ref": "#/components/schemas/UserSchema"
      },
...
    }
  }
}